requirements.txt
openai>=1.0.0
pdfminer.six>=20221105
Pillow>=10.0.0
pytesseract>=0.3.10
python-dotenv>=1.0.0 